backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.3
numpy==2.2.1
ta==0.11.0
ccxt==4.5.43
supabase==2.28.2
pydantic-settings==2.7.1
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
